-реализованы репозитории friend, usertag, Post

+++ Cogni/Authentication/Abstractions/ITokenService.cs
@@ -4,7 +4,7 @@ namespace Cogni.Authentication.Abstractions
 {
     public interface ITokenService
     {
-        string GenerateAccessToken(long id, string role);
+        string GenerateAccessToken(int id, string role);
         string GenerateRefreshToken();
         ClaimsPrincipal GetPrincipalFromExpiredToken(string token);
         DateTime GetRefreshTokenExpireTime();
